src/python/pyproject.toml
[project]
name = "whatsapp-test-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.2.10",
    "openai>=1.82.0",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "boto3 (>=1.38.32,<2.0.0)",
    "email-validator>=2.2.0",
    "pycountry>=24.6.1",
    "pydantic-extra-types>=2.10.5",
    "azure-storage-blob>=12.25.1",
    "pandas>=2.3.0",
    "openpyxl>=3.1.5",
    "shortuuid>=1.0.13",
    "redis>=6.2.0",
]
